Sync ms3 with myosuite4/mjx (squash)
Squash-syncs ms3 to the current state of the Vittorio-Caggiano/myosuite4
fork's mjx branch (d82a2d7f). Picks up mjx commits since the prior sync
point (270eb096):

- feat: migrate myo_sim from a git submodule to the myo-sim PyPI package,
  plus follow-up fixes for path resolution (models/ subdirectory, bimanual
  and tabletennis simhive/myo_sim paths, stale error messages/save paths).
- chore: remove legacy myosuite_init, test_myoapi, myo_model simhive, and
  empty .gitmodules.
- style: pre-commit auto-fixes (pre-push lint pass).
- fix: calibrate myo_sim-pip hand/arm coordinate frame, fix reorient_sar
  obs parity.
- remove placeholder *Modular-v0 challenge registrations, document task
  workflow.
- refactor: dedupe shared mimic infra and remove unused wrapper files.
- fix: pin MJX jax/brax floor, guard macOS MUJOCO_GL, add registry sweep
  test (#97).
