+++ myosuite/integrations/musclemimic/fullbody_eval_cli.py
@@ -366,7 +366,7 @@ def _mjlab_onnx_main(
         compile_mimic_fullbody_mjmodel,
         default_mimic_fullbody_config,
     )
-    from myosuite.integrations.musclemimic.trajectory_io import (
+    from myosuite.core.trajectory_io import (
         load_motion_clip,
         resolve_motion_path,
     )
