--- myosuite/integrations/musclemimic/trajectory_io.py
@@ -1,23 +1,0 @@
-# Copyright (c) MyoSuite Authors. All rights reserved.
-#
-# This source code is licensed under the Apache 2 license found in the
-# LICENSE file in the root directory of this source tree.
-
-"""Compatibility wrapper around core trajectory IO helpers."""
-
-from __future__ import annotations
-
-from myosuite.core.trajectory_io import (
-    MotionClip,
-    expand_motion_clip_to_model,
-    load_motion_clip,
-    resolve_motion_path,
-)
-
-
-__all__ = [
-    "MotionClip",
-    "expand_motion_clip_to_model",
-    "load_motion_clip",
-    "resolve_motion_path",
-]
