--- myosuite/utils/mjlab_play.py
@@ -1,8 +1,0 @@
-# Copyright (c) MyoSuite Authors. All rights reserved.
-#
-# This source code is licensed under the Apache 2 license found in the
-# LICENSE file in the root directory of this source tree.
-
-"""Compatibility wrapper for the mjlab play entrypoint."""
-
-from myosuite.viz.mjlab_play import *  # noqa: F403
