+++ pyproject.toml
@@ -65,8 +65,12 @@ mjlab = [
     "mujoco-warp>=3.7.0; python_version >= '3.10' and python_version < '3.14' and sys_platform != 'win32'",
 ]
 mjx = [
-    "jax>=0.4.20; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
-    "jaxlib>=0.4.20; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
+    # brax's PPO trainer calls jax.device_put_replicated, removed in jax>=0.5.3 —
+    # cap the floor below that line until brax drops the call (verified against
+    # brax 0.10-0.14 series). Without this, an unpinned resolve can silently
+    # break MJX training on any platform, not just the macOS exclusion below.
+    "jax>=0.4.20,<0.5.3; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
+    "jaxlib>=0.4.20,<0.5.3; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
     "brax>=0.10.0; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
     "flax>=0.8.0; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
     "optax>=0.1.7; python_version >= '3.10' and sys_platform != 'win32' and sys_platform != 'darwin'",
@@ -83,7 +87,7 @@ mjx = [
 # Install with: uv sync --extra mjx-cuda
 mjx-cuda = [
     "myosuite[mjx]",
-    "jax[cuda12]>=0.4.30; python_version >= '3.10' and sys_platform != 'darwin'",
+    "jax[cuda12]>=0.4.30,<0.5.3; python_version >= '3.10' and sys_platform != 'darwin'",
 ]
 docs = [
     "sphinx>=7.0",
